requirements.txt
asgiref==3.6.0
Django==4.2.1
psycopg2==2.9.6
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
django-cors-headers==4.0.0